requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
qdrant-client==1.13.3
sentence-transformers==2.7.0
transformers==4.46.0
tokenizers==0.20.0
torch==2.10.0
accelerate>=0.26.0
pymorphy3==2.0.5
nltk==3.8.1
beautifulsoup4==4.12.2
pdfplumber==0.10.3
openpyxl==3.1.2
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
datasets==2.14.0
huggingface-hub==0.26.0
structlog==25.5.0
python-dotenv==1.0.0
evaluate==0.4.0
pyarrow==14.0.1
bert-score==0.3.13
gradio==4.44.0
seaborn==0.13.2
jupyter